requirements.txt
# Tibia Pixel Bot

mss>=9.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pyobjc-framework-Quartz>=9.0.0
pytesseract>=0.3.10
